backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
pillow==10.1.0
django-filter==23.3
openpyxl==3.1.2
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
